reveals flaw in design
unless the monad contains a function we don’t get any means to get at the contents of the monad in the instance functions for the AbstractSelection? right?
so next step is to bring back the Monad (structure -> Tuple a structure) as in Phil’s example and the FakeSelection. But at the same time preserving the typed nature of the D3 Selection