reimplement D3Monad and change dataBind
unfortunately to get this to typecheck i couldn’t get the data being bound to carry same types as the Monad - not sure if this is a limitation of expressiveness of type system or (much more likely) limitation of my ability to express such constraints in the file system. It does sort of seem like something that might be doable in Idris?