Attributes somewhat working…
since i’m going to roll Style and Classed attributes into one single list of attributes and since they take additional parameter, next step is to build support for that in.

NB the way this is actually working ATM the JS depends on the structure of the PS data, ie for the Attr containing the Exists. Plan to revisit and see if this can be made clean and safe again in the future.