this is what Transition would be like…
…if it could be a separate interface…but then we wouldn’t be able to chain from Selection to Transition cause they’d be different monads, different interpreters etc etc

So they’re going to have to be one monad/interpreter and thus we will lose any type safety related to ordering of scripts. This seems very unsatisfactory but don’t currently know of a better way.

This file will accordingly be deleted again in the next commit and the “methods” rolled into “Selection”